[Refactor] Move tablet writer workflow to data_workflows
