+++ be/CMakeLists.txt
@@ -542,7 +542,6 @@ set(STARROCKS_LINK_LIBS
     Webserver
     Tools
     AgentServer
-    DataWorkflows
     ${WL_START_GROUP}
     Connector
     # Legacy connector registration pulls concrete connector objects that
@@ -553,6 +552,7 @@ set(STARROCKS_LINK_LIBS
     FileSystem
     Formats
     Runtime
+    DataWorkflows
     Udf
     Util
     Script
