+++ be/src/data_workflows/load/tablet_writer/lake_tablets_channel.cpp
@@ -33,16 +33,16 @@
 #include "common/system/backend_options.h"
 #include "common/util/stack_trace_mutex.h"
 #include "common/util/table_metrics.h"
+#include "data_workflows/load/tablet_writer/load_channel.h"
+#include "data_workflows/load/tablet_writer/tablets_channel.h"
 #include "fs/bundle_file.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "gutil/macros.h"
 #include "runtime/closure_guard.h"
 #include "runtime/descriptors.h"
-#include "runtime/load_channel.h"
 #include "runtime/mem_pool.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/runtime_metrics.h"
-#include "runtime/tablets_channel.h"
 #include "serde/protobuf_serde.h"
 #include "storage/lake/async_delta_writer.h"
 #include "storage/lake/delta_writer.h"
