+++ be/src/data_workflows/load/tablet_writer/load_channel.cpp
@@ -12,27 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/be/src/runtime/load_channel.cpp
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-#include "runtime/load_channel.h"
+#include "data_workflows/load/tablet_writer/load_channel.h"
 
 #include <memory>
 
@@ -45,14 +25,14 @@
 #include "common/runtime_profile.h"
 #include "common/tracer.h"
 #include "common/util/thrift_util.h"
+#include "data_workflows/load/tablet_writer/lake_tablets_channel.h"
+#include "data_workflows/load/tablet_writer/load_channel_mgr.h"
+#include "data_workflows/load/tablet_writer/local_tablets_channel.h"
 #include "fmt/format.h"
 #include "runtime/closure_guard.h"
 #include "runtime/current_thread.h"
 #include "runtime/descriptors.h"
 #include "runtime/diagnose_daemon.h"
-#include "runtime/lake_tablets_channel.h"
-#include "runtime/load_channel_mgr.h"
-#include "runtime/local_tablets_channel.h"
 #include "runtime/mem_tracker.h"
 #include "runtime/runtime_metrics.h"
