+++ be/src/service/service_be/config_update_hooks.cpp
@@ -46,11 +46,11 @@
 #include "common/util/bthreads/executor.h"
 #include "compute_env/workgroup/scan_executor.h"
 #include "compute_env/workgroup/work_group_manager.h"
+#include "data_workflows/load/tablet_writer/load_channel_mgr.h"
 #include "runtime/batch_write/batch_write_mgr.h"
 #include "runtime/batch_write/txn_state_cache.h"
 #include "runtime/env/global_env.h"
 #include "runtime/exec_env.h"
-#include "runtime/load_channel_mgr.h"
 #include "service/core_dump_resource_releaser.h"
 #include "storage/compaction_manager.h"
 #include "storage/index/vector/vector_index_cache.h"
