+++ be/src/service/service_be/internal_service.cpp
@@ -39,13 +39,13 @@
 #include "compute_env/data_stream/data_stream_mgr.h"
 #include "compute_env/result/buffer_control_block.h"
 #include "compute_env/result/result_buffer_mgr.h"
+#include "data_workflows/load/tablet_writer/load_channel_mgr.h"
 #include "exec/pipeline/fragment_context.h"
 #include "gen_cpp/BackendService.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/closure_guard.h"
 #include "runtime/exec_env.h"
 #include "runtime/fragment_mgr.h"
-#include "runtime/load_channel_mgr.h"
 #include "runtime/routine_load/routine_load_task_executor.h"
 #include "runtime/runtime_filter_worker.h"
 #include "storage/local_tablet_reader.h"
