+++ be/src/service/service_be/lake_service.cpp
@@ -34,6 +34,7 @@
 #include "common/system/cpu_info.h"
 #include "common/thread/thread.h"
 #include "common/thread/threadpool.h"
+#include "data_workflows/load/tablet_writer/load_channel_mgr.h"
 #include "exec/write_combined_txn_log.h"
 #include "fs/fs_factory.h"
 #include "fs/fs_util.h"
@@ -41,7 +42,6 @@
 #include "gutil/strings/join.h"
 #include "runtime/exec_env.h"
 #include "runtime/lake_snapshot_loader.h"
-#include "runtime/load_channel_mgr.h"
 #include "storage/lake/compaction_policy.h"
 #include "storage/lake/compaction_scheduler.h"
 #include "storage/lake/compaction_task.h"
