+++ be/test/data_workflows/CMakeLists.txt
@@ -19,6 +19,10 @@ endif()
 
 set(DATA_WORKFLOWS_TEST_FILES
     ${BE_TEST_MACOS_LINK_STUBS}
+    load/tablet_writer/lake_tablets_channel_test.cpp
+    load/tablet_writer/load_channel_mgr_test.cpp
+    load/tablet_writer/load_channel_test.cpp
+    load/tablet_writer/local_tablets_channel_test.cpp
     push_handler_test.cpp
 )
