+++ be/test/data_workflows/load/tablet_writer/local_tablets_channel_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "runtime/local_tablets_channel.h"
+#include "data_workflows/load/tablet_writer/local_tablets_channel.h"
 
 #include <fmt/format.h>
 #include <gtest/gtest.h>
@@ -29,12 +29,12 @@
 #include "common/config_ingest_fwd.h"
 #include "common/logging.h"
 #include "common/runtime_profile.h"
+#include "data_workflows/load/tablet_writer/load_channel.h"
+#include "data_workflows/load/tablet_writer/load_channel_mgr.h"
 #include "gen_cpp/internal_service.pb.h"
 #include "gutil/walltime.h"
 #include "platform/platform_env.h"
 #include "runtime/exec_env.h"
-#include "runtime/load_channel.h"
-#include "runtime/load_channel_mgr.h"
 #include "runtime/mem_tracker.h"
 #include "serde/protobuf_serde.h"
 #include "storage/chunk_helper.h"
