+++ be/test/service/lake_service_test.cpp
@@ -39,11 +39,11 @@
 #include "column/fixed_length_column.h"
 #include "column/nullable_column.h"
 #include "common/config_lake_fwd.h"
+#include "data_workflows/load/tablet_writer/load_channel_mgr.h"
 #include "fs/fs_util.h"
 #include "gutil/strings/util.h"
 #include "gutil/walltime.h"
 #include "runtime/exec_env.h"
-#include "runtime/load_channel_mgr.h"
 #include "service/brpc_service_test_util.h"
 #include "storage/chunk_helper.h"
 #include "storage/del_vector.h"
